requirements.txt
paho-mqtt==1.5.1
peewee==3.14.4
python-dotenv==0.18.0
